fix: playback fixes
