+++ ui/contexts/expandable.js
@@ -1,0 +1,4 @@
+// @flow
+import React from 'react';
+
+export const ExpandableContext = React.createContext<any>();
